Refactored consolidators, PR response
